Split out the depth stuff from the file, move to GPU/Common
